Revert "Integrate the Irregexp Regular Expression Engine."
This reverts commit https://code.google.com/p/dart/source/detail?r=41949

TBR=fschneider@google.com
BUG=

Review URL: https://codereview.chromium.org//754383002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@41950 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/unibrow-inl.h
@@ -7,12 +7,14 @@
 
 #include "vm/unibrow.h"
 
-#include "platform/assert.h"
+// SNIP
 
 namespace unibrow {
 
-template <class T, intptr_t s> intptr_t Mapping<T, s>::get(
-    int32_t c, int32_t n, int32_t* result) {
+// SNIP
+
+template <class T, int s> int Mapping<T, s>::get(uchar c, uchar n,
+    uchar* result) {
   CacheEntry entry = entries_[c & kMask];
   if (entry.code_point_ == c) {
     if (entry.offset_ == 0) {
@@ -26,10 +28,10 @@ template <class T, intptr_t s> intptr_t Mapping<T, s>::get(
   }
 }
 
-template <class T, intptr_t s> intptr_t Mapping<T, s>::CalculateValue(
-    int32_t c, int32_t n, int32_t* result) {
+template <class T, int s> int Mapping<T, s>::CalculateValue(uchar c, uchar n,
+    uchar* result) {
   bool allow_caching = true;
-  intptr_t length = T::Convert(c, n, result, &allow_caching);
+  int length = T::Convert(c, n, result, &allow_caching);
   if (allow_caching) {
     if (length == 1) {
       entries_[c & kMask] = CacheEntry(c, result[0] - c);
@@ -42,6 +44,8 @@ template <class T, intptr_t s> intptr_t Mapping<T, s>::CalculateValue(
     return length;
   }
 }
+
+// SNIP
 
 }  // namespace unibrow
